requirements.txt
Django==4.0.2
psycopg2==2.9.3
djangorestframework==3.13.1
pandas==1.4.1
numpy==1.22.2
django-cors-headers==3.11.0
djangorestframework-simplejwt==5.0.0